fixed 1st collector with working authentication

+++ kuhako/src/main/java/com/capstone/kuhako/services/CollectorModuleServices/CollectAllDuePaymentsServiceImpl.java
@@ -1,8 +1,10 @@
 package com.capstone.kuhako.services.CollectorModuleServices;
 
 
+import com.capstone.kuhako.models.Collector;
 import com.capstone.kuhako.models.CollectorModules.CollectAllDuePayments;
 import com.capstone.kuhako.repositories.CollectorModuleRepository.CollectAllDuePaymentsRepository;
+import com.capstone.kuhako.repositories.CollectorRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -13,32 +15,51 @@ public class CollectAllDuePaymentsServiceImpl implements CollectAllDuePaymentsSe
     @Autowired
     private CollectAllDuePaymentsRepository collectAllDuePaymentsRepository;
 
+    @Autowired
+    private CollectorRepository collectorRepository;
 
-    public void createCollectAllDuePayments(CollectAllDuePayments collectAllDuePayments){
-//      Collector author = collectorRepository.findByCollectorname(jwtToken.getCollectornameFromToken(stringToken));
-        CollectAllDuePayments newCollectAllDuePayments = new CollectAllDuePayments();
-        newCollectAllDuePayments.setCollector(collectAllDuePayments.getCollector());
-        newCollectAllDuePayments.setCollectionDate(collectAllDuePayments.getCollectionDate());
-        newCollectAllDuePayments.setRequiredCollectibles(collectAllDuePayments.getRequiredCollectibles());
-        newCollectAllDuePayments.setItemCollectible(collectAllDuePayments.getItemCollectible());
-        newCollectAllDuePayments.setTransactionProof(collectAllDuePayments.getTransactionProof());
-        collectAllDuePaymentsRepository.save(newCollectAllDuePayments);
+    public void createCollectAllDuePayments(Long collectorId, CollectAllDuePayments collectAllDuePayments){
+        Collector collector = collectorRepository.findById(collectorId).orElse(null);
+        if (collector != null) {
+            collectAllDuePayments.setCollector(collector);
+            collectAllDuePaymentsRepository.save(collectAllDuePayments);
+        }
     }
+
     public Iterable<CollectAllDuePayments> getCollectAllDuePayments(){
         return collectAllDuePaymentsRepository.findAll();
     }
-    public ResponseEntity deleteCollectAllDuePayments(Long id){
-        collectAllDuePaymentsRepository.deleteById(id);
-        return new ResponseEntity<>("CollectAllDuePayments Deleted successfully", HttpStatus.OK);
+
+    public Iterable<CollectAllDuePayments> getCollectAllDuePaymentsByClientId(Long collectorId){
+        return collectAllDuePaymentsRepository.findCollectAllDuePaymentsByCollectorId(collectorId);
     }
-    public ResponseEntity updateCollectAllDuePayments(Long id, CollectAllDuePayments collectAllDuePayments){
-        CollectAllDuePayments collectAllDuePaymentsForUpdate = collectAllDuePaymentsRepository.findById(id).get();
-        collectAllDuePaymentsForUpdate.setCollector(collectAllDuePayments.getCollector());
-        collectAllDuePaymentsForUpdate.setCollectionDate(collectAllDuePayments.getCollectionDate());
-        collectAllDuePaymentsForUpdate.setRequiredCollectibles(collectAllDuePayments.getRequiredCollectibles());
-        collectAllDuePaymentsForUpdate.setItemCollectible(collectAllDuePayments.getItemCollectible());
-        collectAllDuePaymentsForUpdate.setTransactionProof(collectAllDuePayments.getTransactionProof());
-        collectAllDuePaymentsRepository.save(collectAllDuePaymentsForUpdate);
-        return new ResponseEntity("CollectAllDuePayments updated successfully", HttpStatus.OK);
+
+    public ResponseEntity deleteCollectAllDuePayments(Long collectorId, Long id){
+        CollectAllDuePayments collectAllDuePaymentsToDelete = collectAllDuePaymentsRepository.findById(id).orElse(null);
+
+        if (collectAllDuePaymentsToDelete != null && collectAllDuePaymentsToDelete.getCollector().getCollector_id().equals(collectorId)) {
+            collectAllDuePaymentsRepository.deleteById(id);
+            return new ResponseEntity<>("CollectAllDuePayments Deleted successfully", HttpStatus.OK);
+        } else if (collectAllDuePaymentsToDelete != null) {
+            return new ResponseEntity<>("CollectAllDuePayments does not belong to the specified collector", HttpStatus.NOT_FOUND);
+        } else {
+            return new ResponseEntity<>("CollectAllDuePayments not found", HttpStatus.NOT_FOUND);
+        }
+
+    }
+    public ResponseEntity updateCollectAllDuePayments(Long collectorId, Long id, CollectAllDuePayments collectAllDuePayments){
+        CollectAllDuePayments collectAllDuePaymentsForUpdate = collectAllDuePaymentsRepository.findById(id).orElse(null);
+        if (collectAllDuePaymentsForUpdate != null && collectAllDuePaymentsForUpdate.getCollector().getCollector_id().equals(collectorId)) {
+            collectAllDuePaymentsForUpdate.setCollectionDate(collectAllDuePayments.getCollectionDate());
+            collectAllDuePaymentsForUpdate.setRequiredCollectibles(collectAllDuePayments.getRequiredCollectibles());
+            collectAllDuePaymentsForUpdate.setItemCollectible(collectAllDuePayments.getItemCollectible());
+            collectAllDuePaymentsForUpdate.setTransactionProof(collectAllDuePayments.getTransactionProof());
+            collectAllDuePaymentsRepository.save(collectAllDuePaymentsForUpdate);
+            return new ResponseEntity<>("CollectAllDuePayments Updated successfully", HttpStatus.OK);
+        } else if (collectAllDuePaymentsForUpdate != null) {
+            return new ResponseEntity<>("CollectAllDuePayments does not belong to the specified collector", HttpStatus.NOT_FOUND);
+        } else {
+            return new ResponseEntity<>("CollectAllDuePayments not found",HttpStatus.NOT_FOUND);
+        }
     }
 }
